requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
astroid==2.4.0
autopep8==1.5.5
backcall==0.2.0
cachetools==4.2.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
dialogflow==1.1.0
distlib==0.3.1
et-xmlfile==1.0.1
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
freetype-py==2.2.0
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.26.0
google-auth==1.27.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.3
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
MouseInfo==0.1.3
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.9
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.15.1
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydot==1.4.1
PyGetWindow==0.0.9
Pygments==2.7.1
pylint==2.5.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.3
PyTweening==1.0.3
pytz==2020.1
pyzmq==19.0.2
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.1
sgmllib3k==1.0.0
six==1.14.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.2.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
